Add packages.R to centralize packages
This commit adds the new file packages.R. I found all packages
with the terminal command at the top of packages.R. An alternative
to find packages is env::dependencies(), but it seems to not find
packages on GitHub.

The next step is to remove all calls to library(), to keep the code
clean. Instead, you can now `source("packages.R")`.

Another option is to add a DESCRIPTION file and list the packages
under the field "Imports". That is nice because it brings features
from R packages; but for now this commit keeps it simple.