Add integration test: integration-test.Rmd
This commit adds an integration tests. It is an rmarkdown file
that runs the entire PACTA_analysis, based on Jacob's guide:
https://bit.ly/2RCRJn7

If the file renders successfully, then the project works as
expected. If the file fails to render, then there is a problem we
need to fix.

AFAIK, this is the only test.
TODO: Add finer grained, unit tests.