Merge branch 'master' into current_web_functionality
This reverts commit bb96ed06666ffc887aa98ce685b65302ea475942.

This commit merges current_web_functionality into master. I did
this before and pushed it to GitHub but then I realized that was
a bad idea because this branch was the default branch. So I
reverted that merge commit, then I changed the default branch to
master, and now I'm reverting again to push the merge commit, now
as a pull request.