print git status of all needed repos before run
This should add some output at the top of running the web tool scripts to verify the status of all the need repos. This should help debugging, and help determine differences between runs at different time and/or on different machines.

Does require that the system running it already has `git2r` installed.