Setup environment to test-r-scripts
Closes #285
Replaces #291 #292 and #293 combined.All three PRs touched the
same file and are related in that together they create the
environment for test-r-scripts. This combined PR makes it
easier to review the relationship between all changes, and
avoids merge conflicts. All in all, I expect the reiew to
go more smoothly.