run-web-tool now shows who and when commited last (#301)
It's often useful to know who authored the latest commit and when. This PR adds such information:

➜  PACTA_analysis git:(who-commited-when) ./bin/show-upstream-and-status-at ../PACTA_analysis
PACTA_analysis
Mauro Lepore commited 31 seconds ago:
who-commited-when b5fe4c5 [origin/who-commited-when] Show who commited when
On branch who-commited-when
Your branch is up to date with 'origin/who-commited-when'.

nothing to commit, working tree clean

* Show who commited when

* Style