Assert input to green_brown() has length 1 (#317)
Relates to 2DegreesInvesting/create_interactive_report#177

I see two functions with the name green_brown() -- one in this repository and the other one in create_interactive_report / -- and they are not identical. The version in create_interactive_report/ seems more developed and the web tool seems to be using the version from create_interactive_report/.

Just in case the version in this repo gets called somewhere, I added an assertion to detect the problem that 2DegreesInvesting/create_interactive_report#177 fixed.