also export invalidsecurities as CSV (#428)
closes #427

I'm not thrilled about this PR. I don't fully understand what's being exported here or why. I don't know why it was originally exported as JSON and now it needs to be CSV. I don't know if we should be doing the same thing with the other files that are being exported here also. I don't know if we should keep exporting JSON files and also export CSV files, or if we should pick one. That all being said, this resolves #427 with minimal changes so at least we'll be able to track how/when/why this was done.