Make the good-bahavior test a bit more informative
Testing for the class is a good start but now I would like to
capture a bit more info -- e.g. that with variable values of
`scen_alloc_wt_tech_prod` the `tdm_*` columns do not have NaN.

I could be more specific but for not a snapshot seems appropriate.