Add infrastructure for HTML user error reports
This commit adds a lot to solve a relatively small problem, but opens up
a better line of communication with users in the future.

Generally, this adds `log_user_errors()` and `prepare_error_report()`,
which  allow us to identify user-level errors, which they can fix
before running a new analysis.

Specifically, this checks that the `port_holdings_date` in the
portfolio parameters file is a scalar, not an array (which can happen if
the user groups portfolios with different holding dates.)