read in config values as strings
These values read in from the portfolio config file are always expected to be a string, even if they are a numerical string (e.g. investor_name = "24"). This is a safe guard to make sure that happens when they are read. This should prevent 2DegreesInvesting/create_interactive_report/issues/268 from ever happening.